fastapi_service/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
pillow
qrcode[pil]
pytest
requests
